Move component to Stats dir
