+++ src/components/Stats/index.ts
@@ -1,2 +1,2 @@
 export { default as Stats } from './Stats'
-export { AttackStat, DefenseStat, GoldStat } from './StatsComponents'
+export { AttackStat, DefenseStat, GoldStat, Bar } from './StatsComponents'
