--- src/components/StatsBar/index.ts
@@ -1,1 +1,0 @@
-export { default as StatsBar } from './StatsBar'
